requirements.txt
torch
torchvision
Pillow
numpy
matplotlib
opencv-python
easyocr